Generate a view of the sign suitable for printing, and update the stylesheet so that it will present nicely on print media. Only handles the A4 case at the moment, but does cope with portrait and landscape.
Some care is needed to make sure that the stylesheets work for preview
and print. Basic approach is to set the base font size in `vw` for
preview, and `pt` for print, then use relative lengths, like em and rem,
to size the elements on the page.

Still needs some fine-tuning, and not tested on browsers other than
Chrome yet. #10