Update the process for uploading a new image for the bathing water, in line with change to put all of the options on one screen. There are two key changes here:
- images are now keyed by the bathing-water-controller name from the bw,
not the name given by the user. This means that we can present the
cached logo image as an option, using only the EUBWID as a parameter
(from which we can get the controller name as given in the data). The
images keys on S3 now use the controller name from the data.

- the 'upload a new image' workflow is now in-line with the options
form. This takes away the ability to preview the uploaded image, but
makes the overall workflow much simpler. The user can always come back
to the options screen to upload a different image.