Update to allow for long pollution source descriptions, issue #116. I had missed out adding the measuresSummary statement from the profile where it is provided. In some bathing waters, this makes the pollution sources statement too long to fit in the space available. The agreed solution is to:
- swap the location of the sampling point map and the web-info (QR code)
sections
- in the case of a very long pollution sources statement, hide the map
and allow the content to flow into the space used by the map.

The correct way to solve this would be to use CSS grid, but this is too
much change this close to launch. Instead, we fake it by tweaking the
CSS classes. The downside of this is that we have to lose the grey
border around the pollution sources statement, in the case that it's
very long.

We should come back in the future and revise this to use CSS grid
properly.